refactor/RSS-ECOMM-4_07: replace product short info into separate component
