+++ src/pages/Product/components/ProductShortInfo/ProductShortInfo.module.scss
@@ -1,0 +1,50 @@
+.product-page-title {
+  font-weight: 900;
+  font-size: 30px;
+  margin-bottom: 35px;
+}
+
+.product-page-info {
+  font-size: 24px;
+  font-weight: 900;
+  margin-bottom: 5px;
+  padding-top: 5px;
+}
+
+.product-attributes-container {
+  width: 250px;
+}
+
+.product-attributes {
+  font-weight: 700;
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  font-size: 22px;
+}
+
+.attribute-value,
+.current-price,
+.previous-price {
+  font-weight: 400;
+  margin-left: 15px;
+  width: 50%;
+}
+
+.current-price {
+  color: darkgreen;
+  font-weight: 800;
+  font-size: 28px;
+}
+
+.previous-price {
+  text-decoration: line-through;
+  color: brown;
+  font-weight: 700;
+}
+
+@media (max-width: 950px) {
+  .product-page-title {
+    text-align: center;
+  }
+}
